requirements.txt
discord.py==1.7.3
requests==2.25.1
beautifulsoup4==4.9.3
yt-dlp==2021.12.1
python-dotenv==0.19.0
